Add main entrypoint file frame
Sound Defender presentation

We used this file to hide the domain name during the demo. It is using frame and simply includes URL to the website it is deployed to.

TODO changed the src to the actual local folder and npm run, so we are not dependent on third party